requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=1.5.0
db-dtypes>=1.0.0
flask>=2.3.0
requests>=2.31.0
